% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Returns the current exchange rate from USD to HUF}
\usage{
get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{number of times the function already failed}
}
\value{
number
}
\description{
Returns the current exchange rate from USD to HUF
}
\examples{
get_usdhuf()
}
