% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_usdhufs}
\alias{get_usdhufs}
\title{Returns the exchange rates from USD to HUF in the past 30 days}
\usage{
get_usdhufs(start_date = Sys.Date() - 30, end_date = Sys.Date(), retried = 0)
}
\arguments{
\item{start_date}{date}

\item{end_date}{date}

\item{retried}{number of times the function already failed}
}
\value{
\code{data.table} object
}
\description{
Returns the exchange rates from USD to HUF in the past 30 days
}
