% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Returns the current exchange rate between any currencies}
\usage{
get_exchange_rates(
  base_currency,
  symbol_currency,
  start_date,
  end_date,
  retried = 0
)
}
\arguments{
\item{start_date}{date}

\item{end_date}{date}

\item{retried}{number of times the function already failed}
}
\value{
\code{data.table} object
}
\description{
Returns the current exchange rate between any currencies
}
